Move insert-column into Table
